.history/requirements_20260410144006.txt
fastapi>=0.111.0,<1.0.0
uvicorn[standard]>=0.30.0,<1.0.0
pydantic-settings>=2.3.0,<3.0.0
python-dotenv>=1.0.1,<2.0.0
langchain>=0.2.0
langchain-openai>=0.1.0
langgraph>=0.2.0
pytest>=8.2.0
httpx>=0.27.0
